Add Support for creation of DIY jobs
Update layouts and added event handling in this iteration.
