--- app/src/main/java/ie/wit/doityourself/DIYActivity.kt
@@ -1,17 +1,0 @@
-package ie.wit.doityourself
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import timber.log.Timber
-import timber.log.Timber.i
-
-class DIYActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_diy)
-
-        Timber.plant(Timber.DebugTree())
-
-        i("DIY Activity Started..")
-    }
-}
