+++ app/src/main/java/ie/wit/doityourself/models/DIYModel.kt
@@ -1,0 +1,3 @@
+package ie.wit.doityourself.models
+
+data class DIYModel(var title: String = "")
